Update doc and method signatures.

+++ src/jdk.incubator.vector/share/classes/jdk/incubator/vector/ByteVector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -3330,7 +3330,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
 
     /**
      * Loads a vector from a {@linkplain MemorySegment memory segment}
-     * starting at an offset into the memory segment buffer.
+     * starting at an offset into the memory segment.
      * Bytes are composed into primitive lane elements according
      * to the specified byte order.
      * The vector is arranged into lanes according to
@@ -3365,8 +3365,8 @@ public abstract class ByteVector extends AbstractVector<Byte> {
     }
 
     /**
-     * Loads a vector from a {@linkplain ByteBuffer byte buffer}
-     * starting at an offset into the byte buffer
+     * Loads a vector from a {@linkplain MemorySegment memory segment}
+     * starting at an offset into the memory segment
      * and using a mask.
      * Lanes where the mask is unset are filled with the default
      * value of {@code byte} (zero).
@@ -3377,7 +3377,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
      * <p>
      * The following pseudocode illustrates the behavior:
      * <pre>{@code
-     * MemorySegment slice = ms.asSlice(offset);
+     * var slice = ms.asSlice(offset);
      * byte[] ar = new byte[species.length()];
      * for (int n = 0; n < ar.length; n++) {
      *     if (m.laneIsSet(n)) {
@@ -3821,7 +3821,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
     @Override
     @ForceInline
     public final
-    void intoMemorySegment(MemorySegment ms, int offset,
+    void intoMemorySegment(MemorySegment ms, long offset,
                            ByteOrder bo,
                            VectorMask<Byte> m) {
         if (m.allTrue()) {
@@ -3831,7 +3831,8 @@ public abstract class ByteVector extends AbstractVector<Byte> {
                 throw new IllegalArgumentException();
             }
             ByteSpecies vsp = vspecies();
-            checkMaskFromIndexSize(offset, vsp, m, 1, (int) ms.byteSize());
+            // @@@ downcast from long to int
+            checkMaskFromIndexSize((int) offset, vsp, m, 1, (int) ms.byteSize());
             maybeSwap(bo).intoMemorySegment0(ms, offset, m);
         }
     }
