+++ src/jdk.incubator.vector/share/classes/jdk/incubator/vector/FloatVector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -3023,7 +3023,7 @@ public abstract class FloatVector extends AbstractVector<Float> {
 
     /**
      * Loads a vector from a {@linkplain MemorySegment memory segment}
-     * starting at an offset into the memory segment buffer.
+     * starting at an offset into the memory segment.
      * Bytes are composed into primitive lane elements according
      * to the specified byte order.
      * The vector is arranged into lanes according to
@@ -3058,8 +3058,8 @@ public abstract class FloatVector extends AbstractVector<Float> {
     }
 
     /**
-     * Loads a vector from a {@linkplain ByteBuffer byte buffer}
-     * starting at an offset into the byte buffer
+     * Loads a vector from a {@linkplain MemorySegment memory segment}
+     * starting at an offset into the memory segment
      * and using a mask.
      * Lanes where the mask is unset are filled with the default
      * value of {@code float} (positive zero).
@@ -3070,7 +3070,7 @@ public abstract class FloatVector extends AbstractVector<Float> {
      * <p>
      * The following pseudocode illustrates the behavior:
      * <pre>{@code
-     * MemorySegment slice = ms.asSlice(offset);
+     * var slice = ms.asSlice(offset);
      * float[] ar = new float[species.length()];
      * for (int n = 0; n < ar.length; n++) {
      *     if (m.laneIsSet(n)) {
@@ -3371,7 +3371,7 @@ public abstract class FloatVector extends AbstractVector<Float> {
     @Override
     @ForceInline
     public final
-    void intoMemorySegment(MemorySegment ms, int offset,
+    void intoMemorySegment(MemorySegment ms, long offset,
                            ByteOrder bo,
                            VectorMask<Float> m) {
         if (m.allTrue()) {
@@ -3381,7 +3381,8 @@ public abstract class FloatVector extends AbstractVector<Float> {
                 throw new IllegalArgumentException();
             }
             FloatSpecies vsp = vspecies();
-            checkMaskFromIndexSize(offset, vsp, m, 4, (int) ms.byteSize());
+            // @@@ downcast from long to int
+            checkMaskFromIndexSize((int) offset, vsp, m, 4, (int) ms.byteSize());
             maybeSwap(bo).intoMemorySegment0(ms, offset, m);
         }
     }
